Inject chapter scanner dependency in to Book
This commit implements depdency injection for the chapter scanner to be
used in Book. In a previous commit when the ovedrive media marker
support was added, Book was making it's own decisions about which
scanner to use, but that opened it up to needing to know a lot about
serverSettings which felt wrong.

Instead, now we will set the chapterScanner property on Book to either
OverdriveMediaMarkers or parsing it directly from MediaFile.

This required converting both parses to classe, of which I decided to
make them static utility classes, as instanced classes wouldn't provide
much value here.