Allow user to override log level
This PR adds an environment variable `LOG_LEVEL` that can be used to
control the application log level.  For example, this would allow you to
view dev logs while in production, should you so choose.

If the env var is not set, behavior should not change from the current
behavior.

Env var can be set with the log level string (case insensitive).  For
example:  `LOG_LEVEL=debug`