Use multiplication symbol instead of letter x
Use the proper multiplication symbol for the playback speed. The order
of the number and symbol is reversed so it is read "times 1.0" instead
of "1.0 times" (or previously "1.0 x"). Between the symbol and the
number, a narrow no-break space (&#8239;) is added for proper spacing.
The classes of some spans are removed for better alignment.