Execute docker build-client step on build host, not target
By default buildkit executes all steps using applications built
for the target platform, if necessary by emulation.

The audiobookshelf web client is not architecture-dependent
and it's build takes a long time.

Update Dockerfile to execute the build-client natively on the build
platform, instead of the target platform.

This drastically reduces docker build time for arm platforms on x86.